pgn_rank.dartを作成してPGNRankを移植
